+++ src/ai_karen_engine/core/prompt_router.py
@@ -1,0 +1,3 @@
+from src.core.prompt_router import PluginWrapper, PromptRouter
+
+__all__ = ["PluginWrapper", "PromptRouter"]
